+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add URL string format factory `s.url` (#16)
 - Add Python `dateutil` support for parsing dates (#5)
 
+### Changed
+- `s.all` and `s.any` create `ValidatorSpec`s now rather than `PredicateSpec`s
+  which yield richer error details from constituent Specs
+
 ### Fixed
 - Guard against `inspect.signature` failures with Python builtins (#18)